In the process to move EXXX to instances of Problems to be able to e2e test them

+++ src/se/fredberg/euler/problem/Problem.java
@@ -2,5 +2,5 @@ package se.fredberg.euler.problem;
 
 public interface Problem {
 
-    Integer solve();
+    long solve();
 }
